Time: 0 ms (100.00%) | Memory: 134.2 MB (83.37%) - LeetSync

+++ 2211-k-radius-subarray-averages/k-radius-subarray-averages.cpp
@@ -11,7 +11,7 @@ public:
             sum += nums[right];
 
             if(right - left + 1 == size){
-                int avg = sum/size;
+                double avg = sum/size;
                 res[left + k]=avg;
 
                 sum -= nums[left];
